finished functional example
Created an example where two jobs are pushed into the system. One job
prints a valued passed in the argument. The other job pushes another job
using a counter and then waits on that counter until to hit zero. The
third task prints the "physical" cpu that it's running on.